update gas calculator generator logic
