+++ x/gashub/keeper/genesis.go
@@ -1,8 +1,6 @@
 package keeper
 
 import (
-	"fmt"
-
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/x/gashub/types"
 )
@@ -16,9 +14,6 @@ func (k Keeper) InitGenesis(ctx sdk.Context, genState *types.GenesisState) {
 	for _, mgh := range genState.GetMsgGasParams() {
 		k.SetMsgGasParams(ctx, mgh)
 	}
-
-	// register gas calculators
-	k.RegisterGasCalculators(ctx)
 }
 
 // ExportGenesis returns a GenesisState for a given context and keeper
@@ -30,32 +25,4 @@ func (k Keeper) ExportGenesis(ctx sdk.Context) *types.GenesisState {
 		mghs = append(mghs, *mgh)
 	}
 	return types.NewGenesisState(params, mghs)
-}
-
-func registerAllGasCalculators(msgGasParamsSet []*types.MsgGasParams) {
-	for _, gasParams := range msgGasParamsSet {
-		err := registerSingleGasCalculator(gasParams)
-		if err != nil {
-			panic(err)
-		}
-	}
-}
-
-func registerSingleGasCalculator(gasParams *types.MsgGasParams) error {
-	msgTypeUrl := gasParams.GetMsgTypeUrl()
-
-	switch gasParams.GasParams.(type) {
-	case *types.MsgGasParams_FixedType:
-		types.RegisterCalculatorGen(msgTypeUrl, types.FixedGasCalculatorGen)
-	case *types.MsgGasParams_GrantType:
-		types.RegisterCalculatorGen(msgTypeUrl, types.MsgGrantGasCalculatorGen)
-	case *types.MsgGasParams_MultiSendType:
-		types.RegisterCalculatorGen(msgTypeUrl, types.MsgMultiSendGasCalculatorGen)
-	case *types.MsgGasParams_GrantAllowanceType:
-		types.RegisterCalculatorGen(msgTypeUrl, types.MsgGrantAllowanceGasCalculatorGen)
-	default:
-		return fmt.Errorf("unknown gas params type")
-	}
-
-	return nil
 }
