+++ x/gashub/keeper/keeper.go
@@ -41,7 +41,35 @@ func (k Keeper) Logger(ctx sdk.Context) log.Logger {
 // GetCodec return codec.Codec object used by the keeper
 func (k Keeper) GetCodec() codec.BinaryCodec { return k.cdc }
 
-func (k Keeper) RegisterGasCalculators(ctx sdk.Context) {
-	mghs := k.GetAllMsgGasParams(ctx)
-	registerAllGasCalculators(mghs)
+// GetAuthority returns the x/gashub module's authority.
+func (k Keeper) GetAuthority() string {
+	return k.authority
+}
+
+// GetParams returns the total set of x/gashub parameters.
+func (k Keeper) GetParams(ctx sdk.Context) (params types.Params) {
+	store := ctx.KVStore(k.storeKey)
+	bz := store.Get(types.ParamsKey)
+	if bz == nil {
+		return params
+	}
+
+	k.cdc.MustUnmarshal(bz, &params)
+	return params
+}
+
+// SetParams sets the x/gashub parameters to the param space.
+func (k Keeper) SetParams(ctx sdk.Context, params types.Params) error {
+	if err := params.Validate(); err != nil {
+		return err
+	}
+
+	store := ctx.KVStore(k.storeKey)
+	bz, err := k.cdc.Marshal(&params)
+	if err != nil {
+		return err
+	}
+	store.Set(types.ParamsKey, bz)
+
+	return nil
 }
