--- x/gashub/keeper/querier.go
@@ -1,35 +1,0 @@
-package keeper
-
-//
-// import (
-// 	abci "github.com/cometbft/cometbft/abci/types"
-//
-// 	"github.com/cosmos/cosmos-sdk/codec"
-// 	sdk "github.com/cosmos/cosmos-sdk/types"
-// 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-// 	"github.com/cosmos/cosmos-sdk/x/gashub/types"
-// )
-//
-// // NewQuerier creates a querier for auth REST endpoints
-// func NewQuerier(k Keeper, legacyQuerierCdc *codec.LegacyAmino) sdk.Querier {
-// 	return func(ctx sdk.Context, path []string, req abci.RequestQuery) ([]byte, error) {
-// 		switch path[0] {
-// 		case types.QueryParams:
-// 			return queryParams(ctx, k, legacyQuerierCdc)
-//
-// 		default:
-// 			return nil, sdkerrors.Wrapf(sdkerrors.ErrUnknownRequest, "unknown query path: %s", path[0])
-// 		}
-// 	}
-// }
-//
-// func queryParams(ctx sdk.Context, k Keeper, legacyQuerierCdc *codec.LegacyAmino) ([]byte, error) {
-// 	params := k.GetParams(ctx)
-//
-// 	res, err := codec.MarshalJSONIndent(legacyQuerierCdc, params)
-// 	if err != nil {
-// 		return nil, sdkerrors.Wrap(sdkerrors.ErrJSONMarshal, err.Error())
-// 	}
-//
-// 	return res, nil
-// }
