+++ x/gashub/keeper/store.go
@@ -7,39 +7,6 @@ import (
 	"github.com/cosmos/cosmos-sdk/x/gashub/types"
 )
 
-// GetAuthority returns the x/gashub module's authority.
-func (k Keeper) GetAuthority() string {
-	return k.authority
-}
-
-// GetParams returns the total set of x/gashub parameters.
-func (k Keeper) GetParams(ctx sdk.Context) (params types.Params) {
-	store := ctx.KVStore(k.storeKey)
-	bz := store.Get(types.ParamsKey)
-	if bz == nil {
-		return params
-	}
-
-	k.cdc.MustUnmarshal(bz, &params)
-	return params
-}
-
-// SetParams sets the x/gashub parameters to the param space.
-func (k Keeper) SetParams(ctx sdk.Context, params types.Params) error {
-	if err := params.Validate(); err != nil {
-		return err
-	}
-
-	store := ctx.KVStore(k.storeKey)
-	bz, err := k.cdc.Marshal(&params)
-	if err != nil {
-		return err
-	}
-	store.Set(types.ParamsKey, bz)
-
-	return nil
-}
-
 // GetMsgGasParams get the MsgGasParams associated with a msg type url
 func (k Keeper) GetMsgGasParams(ctx sdk.Context, msgTypeUrl string) types.MsgGasParams {
 	store := ctx.KVStore(k.storeKey)
