--- x/gashub/types/querier.go
@@ -1,6 +1,0 @@
-package types
-
-// query endpoints supported by the gashub Querier
-const (
-	QueryParams = "params"
-)
